+++ HGV.Tarrasque.Api/Models/HeroHistory.cs
@@ -8,7 +8,7 @@ namespace HGV.Tarrasque.Api.Models
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public float Current { get; set; }
-        public float Previous { get; set; }
+        public double Current { get; set; }
+        public double Previous { get; set; }
     }
 }
